pyproject.toml
[project]
name = "pdf2word"
version = "0.2.0"
description = "Convert PDF to DOCX preserving formatting — supports MS Word, LibreOffice, and PyMuPDF engines"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
keywords = ["pdf", "docx", "word", "converter", "libreoffice"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Operating System :: OS Independent",
    "Topic :: Office/Business",
    "Topic :: Text Processing",
]
dependencies = [
    "pymupdf>=1.24.0",
    "python-docx>=1.1.0",
    "pdf2docx>=0.5.8",
    "pdfplumber>=0.11.0",
    "pytesseract>=0.3.10",
    "opencv-python>=4.9.0",
    "Pillow>=10.0.0",
    "pywin32>=306 ; sys_platform == 'win32'",
]

[project.optional-dependencies]
ocr = ["pytesseract>=0.3.10", "opencv-python>=4.9.0"]

[project.scripts]
pdf2word = "pdf2word.cli:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["pdf2word*"]
